+++ Classes/AdminPanel/TypesInformation.php
@@ -89,11 +89,16 @@ final class TypesInformation implements ModuleInterface, ContentProviderInterfac
     }
 
     /**
-     * @return string[]
+     * @return list<string>
      */
     public function getJavaScriptFiles(): array
     {
-        return [\sprintf('EXT:%s/Resources/Public/JavaScript/AdminPanel/Validate.js', Extension::KEY)];
+        $path = \sprintf('EXT:%s/Resources/Public/JavaScript/AdminPanel/', Extension::KEY);
+
+        return [
+            $path . 'Copy.js',
+            $path . 'Validate.js',
+        ];
     }
 
     /**